requirements.txt
torch
diffusers
transformers
sentencepiece
accelerate
Pillow
imageio[ffmpeg]